Move simplify_via_aff to be near simplify_using_aff, for discoverability
